Move all Python files into the main gui/ directory
We don't need separate inside/ and outside/ directories anymore.
Also merge inside/__init__.py and outside/__init__.py.
