+++ tests/gui/rootpassword.py
