+++ tests/gui/summary.py
