+++ tests/gui/welcome.py
